fix: removed kyryllics and fix nvim bugs

+++ config/tmux/scripts/confirm-window-close.zsh
@@ -11,7 +11,7 @@ if [[ -z "$window_id" ]]; then
 fi
 
 case "$response" in
-  y|н)
+  y)
     tmux kill-window -t "$window_id"
     ;;
 esac
